feat: fix submit and skip btns, change right/wrong answers logger
